Adding `manip_midi` to repo
These are programmes to manipulate MIDI data

It works with the paradigm o source - translate - sink

MIDI sources are from MIDI devices (ports) or files, and is written on the standard out by the source programme

MIDI is translated by programmes that read and write MIDI data on the standard in and out.

MIDI sinks are programmes that read MIDI data on standard in and write it to a device (port) or to a file.